refactor(database): renamed directory Database to Cache. Put all the cache objects inside the namespace called cache
